ksym

Struct KallsymsMapped

pub struct KallsymsMapped<'a> { /* private fields */ }

Mapped kallsyms structure from binary blob

Implementations

impl<'a> KallsymsMapped<'a>

pub fn from_blob( blob: &'a [u8], stext: u64, etext: u64, ) -> Result<Self, &'static str>

Convert binary data into the blob structure and return KallsymsMapped.

Safety

The input blob must be well-formed and valid. Undefined behavior may occur if the blob is malformed.

WARNING

The blob is expected to be page-aligned in memory.

pub fn dump_all_symbols(&self) -> String

Dump all symbols in address order and return as a string. Each line: 16-digit hex address + type + name.

pub fn expand_symbol<'b>( &self, bytes: &[u8], name_buf: &'b mut [u8; 1024], ) -> &'b str

Expand a compressed symbol data into the resulting uncompressed string, if uncompressed string is too long (>= maxlen), it will be truncated, given the offset to where the symbol is in the compressed stream.

pub fn lookup_address<'b>( &self, addr: u64, name_buf: &'b mut [u8; 1024], ) -> Option<(&'b str, u64, u64, char)>

Address → Symbol lookup. Returns (symbol name, symbol size, offset within symbol, type) if found.

See https://elixir.bootlin.com/linux/v6.6/source/kernel/kallsyms.c#L446

pub fn lookup_name(&self, name: &str) -> Option<u64>

Lookup the address for this symbol. Returns 0 if not found.

See https://elixir.bootlin.com/linux/v6.6/source/kernel/kallsyms.c#L265

pub fn lookup_names(&self, name: &str, need_end: bool) -> Option<(usize, usize)>

Symbol name → Address lookup (binary search on name order). It returns the first and last name index (need_end) if there are multiple symbols with the same name. If need_end is false, only the first index is valid.

See https://elixir.bootlin.com/linux/v6.6/source/kernel/kallsyms.c#L208