Crate ksym

A library to parse and query the Linux kernel symbol table. ! It can initialize the symbol table from the embedded symbols (when compiled with assembly feature) or from a string in the format of /proc/kallsyms.

Statics

KSYM

The global kernel symbol table.

Functions

addr_from_symbol

Get the address of the symbol. If not found, return None.

init_kernel_symbols

Initialize the kernel symbol table from a string in the format of /proc/kallsyms

lookup_kallsyms

Return the function information according to the pc address. If not found, return None.