Crate ksym

A library to parse and query the Linux kernel symbol table. ! It can initialize the symbol table from the embedded symbols (when compiled with assembly feature) or from a string in the format of /proc/kallsyms.

Structs

KallsymsMapped

Mapped kallsyms structure from binary blob

Constants

KSYM_NAME_LEN

Maximum length for a symbol name

TOKEN_MARKER

Token delimiter marker